+++ src/settings.rs
@@ -53,7 +53,7 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn set_max_call_levels(&mut self, levels: usize) -> &mut Self {
-        self.limits.max_call_stack_depth = levels;
+        self.limits_set.max_call_stack_depth = levels;
         self
     }
 
@@ -61,7 +61,7 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn max_call_levels(&self) -> usize {
-        self.limits.max_call_stack_depth
+        self.limits_set.max_call_stack_depth
     }
 
     /// Set the maximum number of operations allowed for a script to run to avoid
@@ -69,7 +69,7 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn set_max_operations(&mut self, operations: u64) -> &mut Self {
-        self.limits.max_operations = if operations == u64::MAX {
+        self.limits_set.max_operations = if operations == u64::MAX {
             0
         } else {
             operations
@@ -81,14 +81,14 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn max_operations(&self) -> u64 {
-        self.limits.max_operations
+        self.limits_set.max_operations
     }
 
     /// Set the maximum number of imported modules allowed for a script.
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn set_max_modules(&mut self, modules: usize) -> &mut Self {
-        self.limits.max_modules = modules;
+        self.limits_set.max_modules = modules;
         self
     }
 
@@ -96,7 +96,7 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn max_modules(&self) -> usize {
-        self.limits.max_modules
+        self.limits_set.max_modules
     }
 
     /// Set the depth limits for expressions (0 for unlimited).
@@ -105,18 +105,21 @@ impl Engine {
     pub fn set_max_expr_depths(
         &mut self,
         max_expr_depth: usize,
-        max_function_expr_depth: usize,
+        #[cfg(not(feature = "no_function"))] max_function_expr_depth: usize,
     ) -> &mut Self {
-        self.limits.max_expr_depth = if max_expr_depth == usize::MAX {
+        self.limits_set.max_expr_depth = if max_expr_depth == usize::MAX {
             0
         } else {
             max_expr_depth
         };
-        self.limits.max_function_expr_depth = if max_function_expr_depth == usize::MAX {
-            0
-        } else {
-            max_function_expr_depth
-        };
+        #[cfg(not(feature = "no_function"))]
+        {
+            self.limits_set.max_function_expr_depth = if max_function_expr_depth == usize::MAX {
+                0
+            } else {
+                max_function_expr_depth
+            };
+        }
         self
     }
 
@@ -124,21 +127,22 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn max_expr_depth(&self) -> usize {
-        self.limits.max_expr_depth
+        self.limits_set.max_expr_depth
     }
 
     /// The depth limit for expressions in functions (0 for unlimited).
     #[cfg(not(feature = "unchecked"))]
+    #[cfg(not(feature = "no_function"))]
     #[inline(always)]
     pub fn max_function_expr_depth(&self) -> usize {
-        self.limits.max_function_expr_depth
+        self.limits_set.max_function_expr_depth
     }
 
     /// Set the maximum length of strings (0 for unlimited).
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn set_max_string_size(&mut self, max_size: usize) -> &mut Self {
-        self.limits.max_string_size = if max_size == usize::MAX { 0 } else { max_size };
+        self.limits_set.max_string_size = if max_size == usize::MAX { 0 } else { max_size };
         self
     }
 
@@ -146,7 +150,7 @@ impl Engine {
     #[cfg(not(feature = "unchecked"))]
     #[inline(always)]
     pub fn max_string_size(&self) -> usize {
-        self.limits.max_string_size
+        self.limits_set.max_string_size
     }
 
     /// Set the maximum length of arrays (0 for unlimited).
@@ -154,7 +158,7 @@ impl Engine {
     #[cfg(not(feature = "no_index"))]
     #[inline(always)]
     pub fn set_max_array_size(&mut self, max_size: usize) -> &mut Self {
-        self.limits.max_array_size = if max_size == usize::MAX { 0 } else { max_size };
+        self.limits_set.max_array_size = if max_size == usize::MAX { 0 } else { max_size };
         self
     }
 
@@ -163,7 +167,7 @@ impl Engine {
     #[cfg(not(feature = "no_index"))]
     #[inline(always)]
     pub fn max_array_size(&self) -> usize {
-        self.limits.max_array_size
+        self.limits_set.max_array_size
     }
 
     /// Set the maximum length of object maps (0 for unlimited).
@@ -171,7 +175,7 @@ impl Engine {
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
     pub fn set_max_map_size(&mut self, max_size: usize) -> &mut Self {
-        self.limits.max_map_size = if max_size == usize::MAX { 0 } else { max_size };
+        self.limits_set.max_map_size = if max_size == usize::MAX { 0 } else { max_size };
         self
     }
 
@@ -180,7 +184,7 @@ impl Engine {
     #[cfg(not(feature = "no_object"))]
     #[inline(always)]
     pub fn max_map_size(&self) -> usize {
-        self.limits.max_map_size
+        self.limits_set.max_map_size
     }
 
     /// Set the module resolution service used by the `Engine`.